pyproject.toml
[project]
name = "hangman"
version = "0.1.0"
description = "a modular, extensible codebase in Python using LangChain and LangGraph to test various agents' abilities to maintain a private state"
authors = [
    {name = "Dundalia",email = "davide.baldelli98@gmail.com"}
]
license = {text = "MIT"}
packages = [{include = "hangman", from = "src"}]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "vllm (>=0.5.0)",
    "transformers (>=4.53.3,<5.0.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "accelerate (>=1.9.0,<2.0.0)",
    "smolagents (>=1.20.0,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langgraph (>=0.5.4,<0.6.0)",
    "langchain-anthropic (>=0.3.17,<0.4.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "huggingface-hub (>=0.33.4,<0.34.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "random-word (>=1.0.13,<2.0.0)",
    "anthropic (>=0.59.0,<0.60.0)",
    "diff-match-patch (>=20241021,<20241022)",
    "jupyter (>=1.1.1,<2.0.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "tavily-python (>=0.7.10,<0.8.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "e2b-code-interpreter (>=1.5.2,<2.0.0)",
    "ddgs (>=9.5.2,<10.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "wordfreq (>=3.1.1,<4.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "mem0ai (==1.0.0b0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "litellm (>=1.77.7,<2.0.0)",
    "agentic-memory @ file:///network/scratch/b/baldelld/A-mem-sys",
    "letta-client (>=1.6.0,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "lightmem (>=0.0.0,<0.0.1)",
    "memoryos-pro (>=0.1.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.dependencies]
mem0ai = {allow-prereleases = true}
agentic-memory = {develop = true}
